inventory/inventory/utils/qr_code.py
import os
from PIL import Image
import numpy as np
from .database import IMAGES_DIR

# Mock QR code functionality for development
# In a production environment, install the qrcode, pyzbar, and opencv-python packages

def generate_qr_code(sku):
    """Generate a placeholder QR code for a given SKU"""
    # Create a blank image
    img = Image.new('RGB', (300, 300), color='white')
    
    # Save the image
    qr_path = os.path.join(IMAGES_DIR, f"qr_{sku}.png")
    img.save(qr_path)
    
    print(f"Placeholder QR code created for {sku}")
    return qr_path

def scan_qr_code_from_image(image_array):
    """Mock scanning QR code from an image array"""
    # In a real implementation, this would use pyzbar and opencv
    # For now, return a mock result for testing
    print("QR code scanning simulation (returning mock data)")
    return "LAB-00001"