inventory/requirements.txt
streamlit
pandas
pillow
pytest
black
qrcode
streamlit-cookies